+++ R/zzz.R
@@ -1,7 +1,3 @@
-#' @importFrom lmtp static_binary_on
-#' @importFrom lmtp static_binary_off
-NULL
-
 .onLoad <- function(...) {
 	S7::methods_register()
 }